+++ examples/id.js
@@ -1,3 +1,5 @@
+'use strict'
+
 var ipfsd = require('../index.js')
 
 ipfsd.disposableApi(function (err, ipfs) {
