+++ examples/local.js
@@ -1,3 +1,5 @@
+'use strict'
+
 var ipfsd = require('../index.js')
 
 // opens an api connection to local running ipfs node
