+++ src/node.js
@@ -126,17 +126,17 @@ module.exports = class Node {
 
       this.subprocess = run(this.exec, ['daemon'], {env: this.env})
         .on('error', (err) => {
-          if ((err + '').match('daemon is running')) {
+          if (String(err).match('daemon is running')) {
             // we're good
             done(null, ipfs(conf.Addresses.API))
-          } else if ((err + '').match('non-zero exit code')) {
+          } else if (String(err).match('non-zero exit code')) {
             // ignore when kill -9'd
           } else {
             done(err)
           }
         })
         .on('data', (data) => {
-          const match = (data + '').trim().match(/API server listening on (.*)/)
+          const match = String(data).trim().match(/API server listening on (.*)/)
           if (match) {
             this.apiAddr = match[1]
             const addr = multiaddr(this.apiAddr).nodeAddress()
