--- test/.eslintrc
@@ -1,7 +1,0 @@
-{
-  "globals": {
-    "describe": true,
-    "it": true,
-    "before": true
-  }
-}
